* 'uint64_t' types.
git-svn-id: svn+ssh://multi2sim.org/multi2sim/trunk@487 a6503f54-2359-11de-8ce0-bfd7a313a9c3

+++ src/libgpukernel/gpukernel.c
@@ -30,8 +30,8 @@
 
 struct gk_t *gk;
 
-uint64_t gpu_max_cycles = 0;
-uint64_t gpu_max_inst = 0;
+long long gpu_max_cycles = 0;
+long long gpu_max_inst = 0;
 int gpu_max_kernels = 0;
 
 enum gpu_sim_kind_t gpu_sim_kind = gpu_sim_functional;
